Support 3 element color specifications
BMProperty4D is only used for color values. The getValue method would
reject color specifications with fewer than 4 elements, causing
3-element RGB colors to be rendered as invisible (transparent). Such
color specifications are common in lottie files, so fix the code to
accept them and set the default alpha value to 1, fully opaque.

As a driveby, limit color element values to [0-1], since they will be
passed to QColor which will change to extended color spec on values
outside this range, and that is not wanted here.

Change-Id: I22bb251fc3b159976e965a17c27735ac8fd52ff3
Pick-to: 6.8
Reviewed-by: Kwanghyo Park <kwanghyo.park@qt.io>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>