Fix decoding of shape direction setting
The path direction ("d") setting in a lottie file is read as an int,
and any non-zero value was interpreted as reversed direction. However,
the direction setting is documented to be an enum with values 1 for
Normal, and 3 for Reversed.
Add a function to ensure common and correct interpretation.

Also fix the default painting direction of the ellipse shape: Lottie
by default draws clockwise, while Qt's default is counterclockwise.

Change-Id: Ib1202fd69f409cca5bd158d40a6d7204bc6236dd
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>
(cherry picked from commit 5ce234c32853c68dbf2eb5a0422e805a1db1755b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>